+++ source/direct3d11/TimestampQueryData11.cpp
@@ -31,7 +31,7 @@ namespace Direct3D11
 {
 	TimestampQueryData::TimestampQueryData( const D3D11_QUERY_DATA_TIMESTAMP_DISJOINT& native )
 	{
-		Frequency = static_cast<long>( native.Frequency );
+		Frequency = native.Frequency;
 		IsDisjointed = native.Disjoint != 0;
 	}
